# [SPARK-44649][SQL] Runtime Filter supports passing equivalent creation side expressions