@cloud-fan 2023-09-01
I have no idea how to review the code change here. I understand the general idea but how to map it to the actual code? What does the `def extract` do today and why do you add a new parameter to it?